+++ api/login/repository_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/jinzhu/gorm"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	. "github.com/pass-wall/passwall-server/login"
+	. "github.com/pass-wall/passwall-server/api/login"
 	// "github.com/pass-wall/passwall-server/pkg/config"
 )
