+++ cmd/passwall-server/main_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/gin-gonic/gin"
-	"github.com/pass-wall/passwall-server/login"
+	"github.com/pass-wall/passwall-server/api/login"
 	"github.com/pass-wall/passwall-server/internal/router"
 	"github.com/stretchr/testify/assert"
 )
