+++ util/generate-password.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	"github.com/pass-wall/passwall-server/api/login"
 	"github.com/pass-wall/passwall-server/helper"
-	"github.com/pass-wall/passwall-server/login"
 )
 
 // GeneratePassword generates new password
